requirements.txt
stream-framework==1.3.7
PyGithub==1.43.4
redis==2.10.6
Jinja2==2.10